Add files via upload
Retrieves values from funds in MorningStar via http using the selenium module for Powershell. The Firefox browser must be installed.
The ID of the funds have to be provided in a text file, each in one line.
The output will be written to the current directory as 'fund_list_<date>.csv'.